requirements.txt
asgiref==3.2.3
cffi==1.13.2
cryptography==2.8
Django==3.0.2
django-bootstrap-form==3.4
django-ckeditor==5.8.0
django-crispy-forms==1.8.1
django-filter==2.2.0
django-js-asset==1.2.2
gunicorn==20.0.4
mysqlclient==1.4.6
Pillow==7.0.0
pycparser==2.19
PyMySQL==0.9.3
python-slugify==4.0.0
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
text-unidecode==1.3
